mm-fix-huge-zero-page-accounting-in-smaps-report-fix-2
fs/proc/task_mmu.c:474 smaps_account() warn: should 'size << 12' be a 64 bit type?

Reported-by: Wu Fengguang <fengguang.wu@intel.com>

+++ fs/proc/task_mmu.c
@@ -465,16 +465,15 @@ static void smaps_account(struct mem_size_stats *mss, struct page *page,
 			mss->shared_dirty += size;
 		else
 			mss->shared_clean += size;
-		mss->pss += (size << PSS_SHIFT) / mapcount;
+		mss->pss += ((u64)size << PSS_SHIFT) / mapcount;
 	} else {
 		if (dirty || PageDirty(page))
 			mss->private_dirty += size;
 		else
 			mss->private_clean += size;
-		mss->pss += (size << PSS_SHIFT);
+		mss->pss += (u64)size << PSS_SHIFT;
 	}
 }
-
 
 static void smaps_pte_entry(pte_t *pte, unsigned long addr,
 		struct mm_walk *walk)
